refactor to simplified code and eliminate warnings

+++ src/org/alchemy/core/AlcColorSelector.java
@@ -221,9 +221,9 @@ public class AlcColorSelector extends JDialog implements DocumentListener, AlcCo
      * Set the RGB values based on the current HSB values.
      */
     private void updateRGB() {
-        int rgb = Color.HSBtoRGB((float) hue / 359f,
-                (float) saturation / 99f,
-                (float) brightness / 99f);
+        int rgb = Color.HSBtoRGB(hue / 359f,
+                saturation / 99f,
+                brightness / 99f);
         updateRGB2(rgb);
     }
